Fixing bugs in task 0

+++ alx_book_store.sql
@@ -43,7 +43,7 @@ CREATE TABLE Order_Details(
 	orderdetail_id INT PRIMARY KEY,
 	order_id INT,
 	book_id INT,
-	quantity FLOAT,
+	quantity DOUBLE,
 	FOREIGN KEY (order_id) REFERENCES Orders(order_id),
 	FOREIGN KEY (book_id) REFERENCES Books(book_id)
 );
